modify MAX_NF and temporarily remove the usage of BIGINT in cufinufft source and test files

+++ test/cufinufft1d1_test.cu
@@ -163,7 +163,7 @@ int main(int argc, char* argv[])
 
 	int nt1 = (int)(0.37*N1);  // choose some mode index to check
 	CPX Ft = CPX(0,0), J = IMA*(FLT)iflag;
-	for (BIGINT j=0; j<M; ++j)
+	for (int j=0; j<M; ++j)
 		Ft += c[j] * exp(J*(nt1*x[j]));   // crude direct
 	int it = N1/2+nt1;   // index in complex F as 1d array
 //	printf("[gpu   ] one mode: abs err in F[%ld is %.3g\n",(int)nt1,abs(Ft-fk[it]));
